backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-storage==2.10.0
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
requests==2.31.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
